libretro-common: Clean up dependencies of file_path
This moves a few functions that were defined in file_path_special to libretro-common to clean up the dependency tree.
